Revert "feat: linked graph view with navigation"
This reverts commit 90f8e950faf0567eef7538a3dff4bc5c3db4c431.
