+++ components/Graph/ForceDirectedGraph/ForceDirectedGraph.tsx
@@ -2,7 +2,10 @@ import React, { useEffect, useRef, useState } from "react"
 import {
   ActivityIndicator,
   View,
-  Dimensions
+  Dimensions,
+  Modal,
+  Text,
+  TouchableWithoutFeedback,
 } from "react-native"
 import Svg, { Circle, G, Line, Text as SVGText, Image } from "react-native-svg"
 
@@ -16,8 +19,6 @@ import {
   State,
   GestureHandlerRootView,
 } from "react-native-gesture-handler"
-
-import { black, peach, transparent } from "../../../assets/colors/colors"
 
 import profile_picture_0 from "../../../assets/graph-template-profile-pictures/graph-template-profile-picture-0.png"
 import profile_picture_1 from "../../../assets/graph-template-profile-pictures/graph-template-profile-picture-1.png"
@@ -42,12 +43,10 @@ const SHORT_PRESS_DURATION = 100
 const NODE_HITBOX_SIZE = 20 // Hitbox size of the nodes
 const DEFAULT_NODE_SIZE = 10 // Default size of the nodes
 const DEFAULT_NODE_SIZE_INCREMENT = 2 // Increment in the size of the nodes
-const NODE_HIGHLIGHT_RATIO = 1.3
-const NODE_TEXT_OFFSET = 1.5
 
-const DEFAULT_LINK_COLOR = black // Default color of the links
+const DEFAULT_LINK_COLOR = "black" // Default color of the links
 const DEFAULT_CLICKED_NODE_ID = ""
-const INITIAL_SCALE = 1 // Initial scale of the graph
+
 const MAX_ITERATIONS = 1000 // Maximum number of iterations for the used algorithn
 
 const WIDTH = Dimensions.get("window").width // Width of the screen
@@ -70,9 +69,7 @@ const TOTAL_FRAMES = FPS * (ANIMATION_DURATION / 1000) // Total number of frames
 const ForceDirectedGraph: React.FC<{
   graph: Graph
   constrainedNodeId: string
-  onModalPress: (uid: string) => void
-}> = ({ graph, constrainedNodeId, onModalPress}) => {
-
+}> = ({ graph, constrainedNodeId }) => {
   // States to store the nodes, links, sizes and loading status
   const [nodes, setNodes] = useState<Node[]>([])
   const [links, setLinks] = useState<Link[]>([])
@@ -84,10 +81,12 @@ const ForceDirectedGraph: React.FC<{
   const [clickedNodeID, setClickedNodeID] = useState<string>(
     DEFAULT_CLICKED_NODE_ID,
   ) // Node ID of clicked node
-  const [scale, setScale] = useState(INITIAL_SCALE)
-  const [lastScale, setLastScale] = useState(INITIAL_SCALE) // Add state to keep track of last scale
+  const [scale, setScale] = useState(1)
+  const [lastScale, setLastScale] = useState(1) // Add state to keep track of last scale
 
   const [gestureEnabled, setGestureEnabled] = useState(true) // Add state to keep track of animation start
+
+  const [modalVisible, setModalVisible] = useState(false)
 
   const pressStartRef = useRef(0)
 
@@ -97,35 +96,28 @@ const ForceDirectedGraph: React.FC<{
   // Use effect to update the graph
   useEffect(() => {
     // Get the initial links, nodes and sizes
-    
     const initialLinks = graph.getLinks()
-    const initialSizes = setNodesSizes(initialLinks)
+    const initialNodes = graph.getNodes()
+    const initialSizes = setNodesSizes([...initialLinks])
     // Set the links, nodes and sizes
-    setLinks(initialLinks)
-    if (graph.getInitialized() === false) {
-      setNodes(
-        fruchtermanReingold(
-          graph.getNodes(),
-          initialLinks,
-          constrainedNodeId,
-          WIDTH,
-          HEIGHT,
-          MAX_ITERATIONS,
-        ),
-      )
-      graph.setInitialized(true)
-    }
-    else {
-      setNodes(graph.getNodes())
-    }
+    setLinks([...initialLinks])
+    setNodes(
+      fruchtermanReingold(
+        [...initialNodes],
+        initialLinks,
+        constrainedNodeId,
+        WIDTH,
+        HEIGHT,
+        MAX_ITERATIONS,
+      ),
+    )
     setSizes(initialSizes)
     setLoad(true)
-
   }, [graph, constrainedNodeId])
 
   // If the graph is not loaded, display an activity indicator
   if (!load) {
-    return <ActivityIndicator size="large" color={peach} />
+    return <ActivityIndicator size="large" color="#0000ff" />
   }
 
   // Handle Dragging
@@ -150,6 +142,7 @@ const ForceDirectedGraph: React.FC<{
           y: coordY(node),
         })),
       )
+
       setClickedNodeID(DEFAULT_CLICKED_NODE_ID)
       setTotalOffset({ x: 0, y: 0 })
     }
@@ -187,6 +180,7 @@ const ForceDirectedGraph: React.FC<{
   const nodeZoomIn = (clickedNode: Node) => {
 
     setGestureEnabled(false) // Set animation started to true
+    setModalVisible(true)
 
     let currentFrame = 0
 
@@ -307,13 +301,6 @@ const ForceDirectedGraph: React.FC<{
 
   const CIRCLES = nodes.map((node) => (
     <G key={node.id + "group"}>
-      <Circle
-        key={node.id + "highlight"}
-        cx={coordX(node)}
-        cy={coordY(node)}
-        r={(sizes.get(node.id) ?? DEFAULT_NODE_SIZE) * (graph.getNodeById(node.id)?.selected ? NODE_HIGHLIGHT_RATIO : 1)}
-        fill={peach} 
-      />
       <Image
         key={node.id + "image"}
         x={coordX(node) - (sizes.get(node.id) ?? DEFAULT_NODE_SIZE)}
@@ -330,7 +317,7 @@ const ForceDirectedGraph: React.FC<{
           (sizes.get(node.id) ?? DEFAULT_NODE_SIZE) +
           (NODE_HITBOX_SIZE) / lastScale
         }
-        fill={transparent}
+        fill={"transparent"}
         onPressIn={() => {
           handlePressIn(() => {
             pressStartRef.current = Date.now()
@@ -340,8 +327,6 @@ const ForceDirectedGraph: React.FC<{
         }
         onPressOut={() =>{
           handlePressOut(() => {
-            onModalPress(node.id)
-            setClickedNodeID(DEFAULT_CLICKED_NODE_ID)
             nodeZoomIn(node)
           })
         }
@@ -354,7 +339,7 @@ const ForceDirectedGraph: React.FC<{
         y={
           coordY(node) +
           (sizes.get(node.id) ?? DEFAULT_NODE_SIZE) +
-          NODE_TEXT_OFFSET * DEFAULT_NODE_SIZE
+          DEFAULT_NODE_SIZE
         } // Position below the circle adjust 10 as needed
         textAnchor="middle" // Center the text under the circle
       >
@@ -365,6 +350,48 @@ const ForceDirectedGraph: React.FC<{
 
   return (
     <View style={styles.container}>
+      <Modal
+        animationType="fade"
+        transparent={true}
+        visible={modalVisible}
+        testID="modal"
+      >
+        <TouchableWithoutFeedback
+          onPress={() => {
+            setModalVisible(false)
+            setClickedNodeID(DEFAULT_CLICKED_NODE_ID)
+            setGestureEnabled(true)
+          }}
+          testID="modal-touchable"
+        >
+          <View style={styles.modalContainer}>
+            <TouchableWithoutFeedback>
+              <View style={styles.modalView}>
+                <Svg style={styles.modalProfilePicture}>
+                  <Image
+                    key={clickedNodeID + "modalimage"}
+                    width={80}
+                    height={80}
+                    href={
+                      PROFILE_PICTURES[
+                        parseInt(clickedNodeID) % PROFILE_PICTURES.length
+                      ]
+                    }
+                    onPress={() => {
+                      console.warn("Pressed")
+                    }}
+                    testID="modal-profile-picture"
+                  />
+                </Svg>
+                <Text style={styles.modalProfileName}>
+                  Node ID: {clickedNodeID}
+                </Text>
+              </View>
+            </TouchableWithoutFeedback>
+          </View>
+        </TouchableWithoutFeedback>
+      </Modal>
+
       <GestureHandlerRootView style={styles.container}>
         <PinchGestureHandler
           ref={pinchRef}
@@ -403,7 +430,6 @@ const ForceDirectedGraph: React.FC<{
     </View>
   )
 }
-
 
 export default ForceDirectedGraph
