--- components/Graph/NodeModal/styles.ts
@@ -1,54 +1,0 @@
-import { StyleSheet } from "react-native"
-import { black, lightPeach, peach } from "../../../assets/colors/colors"
-import { globalStyles } from "../../../assets/global/globalStyles"
-
-const styles = StyleSheet.create({
-  container: {
-    alignItems: "center",
-    flex: 1,
-    justifyContent: "center",
-  },
-  modalContainer: {
-    alignItems: "center",
-    flex: 1,
-    justifyContent: "flex-end",
-  },
-  modalProfileName: {
-    fontFamily: globalStyles.text.fontFamily,
-    fontSize: 20,
-    left: 20,
-    position: "absolute",
-    top: 120,
-  },
-  modalProfilePicture: {
-    alignItems: "center",
-    height: 80,
-    justifyContent: "center",
-    left: 20,
-    position: "absolute",
-    top: 20,
-    width: 80,
-  },
-  modalView: {
-    alignItems: "center",
-    backgroundColor: peach,
-    borderColor: lightPeach,
-    borderRadius: 20,
-    borderWidth: 3,
-    elevation: 5,
-    height: "40%",
-    justifyContent: "center",
-    margin: 20,
-    padding: 35,
-    shadowColor: black,
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 4,
-    width: "90%",
-  },
-})
-
-export default styles
