--- screens/Contacts/ContactGraph/mockGraph.tsx
@@ -1,10 +1,0 @@
-import Graph from "../../../components/Graph/Graph"
-
-export const graph = new Graph(
-    ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10"],
-    ["0", "2", "3", "4", "5", "6", "7", "8", "9", "10"],
-    ["1", "1", "1", "1", "1", "1", "2", "2", "3", "3"],
-    [0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5]
-  )
-
-export const constrainedNodeId = "1"
